+++ fourtoon-cookie/src/pages/CharacterSelectPage/ArtworkList.tsx
@@ -1,10 +1,11 @@
 import React, { useMemo } from 'react';
 import { Text, View, FlatList } from 'react-native';
-import CharacterList from '../../../components/character/CharacterList/CharacterList';
-import type { Character, CharacterPaymentType } from '../../../types/character';
+import CharacterList from '../../components/character/CharacterList/CharacterList';
+import type { Character, CharacterPaymentType } from '../../types/character';
 
-import * as S from './ArtworkList.styeld';
-import { useCharacters } from '../../../hooks/server/character';
+import { useCharacters } from '../../hooks/server/character';
+
+import { StyleSheet } from 'react-native';
 
 export interface ArtworkListProps {
 	paymentType: CharacterPaymentType
@@ -59,8 +60,8 @@ interface ArtworkItemProps {
 const ArtworkItem = (props: ArtworkItemProps) => {
     const { artworkTitle, artworkCharacters, ...rest } = props;
     return (
-        <View key={artworkTitle} style={S.styles.artworkContainer}>
-			<Text style={S.styles.artworkTitle}>{artworkTitle}</Text>
+        <View key={artworkTitle} style={styles.artworkContainer}>
+			<Text style={styles.artworkTitle}>{artworkTitle}</Text>
 			<CharacterList
 				characters={artworkCharacters}
 				numColumns={3}
@@ -70,3 +71,25 @@ const ArtworkItem = (props: ArtworkItemProps) => {
 };
 
 export default ArtworkList;
+
+const styles = StyleSheet.create({
+    artworkContainer: {
+        flex: 1,
+        justifyContent: 'flex-start',
+        padding: 0,
+        marginTop: 20,
+        gap: 12,
+        width: "100%",
+        height: "100%",
+    },
+    artworkTitle: {
+        width: "100%",
+        height: 22,
+        fontFamily: 'Pretendard',
+        fontWeight: '600',
+        fontSize: 17,
+        lineHeight: 22,
+        letterSpacing: -0.005,
+        color: '#212121',
+    },
+});
