--- fourtoon-cookie/src/pages/CharacterSelectPage/ArtworkList/ArtworkList.styeld.ts
@@ -1,23 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    artworkContainer: {
-        flex: 1,
-        justifyContent: 'flex-start',
-        padding: 0,
-        marginTop: 20,
-        gap: 12,
-        width: "100%",
-        height: "100%",
-    },
-    artworkTitle: {
-        width: "100%",
-        height: 22,
-        fontFamily: 'Pretendard',
-        fontWeight: '600',
-        fontSize: 17,
-        lineHeight: 22,
-        letterSpacing: -0.005,
-        color: '#212121',
-    },
-});
