--- fourtoon-cookie/src/pages/CharacterSelectPage/CharacterSelectPage.styled.ts
@@ -1,18 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        paddingLeft: 20,
-        paddingRight: 20,
-        position: 'relative',
-        backgroundColor: '#FFFFFF',
-    },
-    separator: {
-        width: "100%",
-        height: 1,
-        backgroundColor: '#F7F7F7',
-        alignSelf: 'stretch',
-        flexGrow: 0,
-    }
-});
