+++ fourtoon-cookie/src/pages/CharacterSelectPage/CharacterSelectPage.tsx
@@ -1,13 +1,12 @@
 import React, { useEffect, useState } from 'react';
-import { View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { CharacterPaymentType } from '../../types/character';
 
-import Header from './Header/Header';
-import TabsLayout from './TabsLayout/TabsLayout';
-import ArtworkList from './ArtworkList/ArtworkList';
+import Header from './Header';
+import TabsLayout from './TabsLayout';
+import ArtworkList from './ArtworkList';
 
-import * as S from './CharacterSelectPage.styled';
 import { useFunctionWithErrorHandling } from '../../hooks/error';
 
 
@@ -21,13 +20,13 @@ const CharacterSelectPage = () => {
     });
 
     return (
-        <SafeAreaView style={S.styles.container}>
+        <SafeAreaView style={styles.container}>
             <Header />
             <TabsLayout
                 selectedPaymentType={selectedPaymentType}
                 onSelectedPaymentTypeChange={handleSelectedPaymentTypeChange}
             />
-            <View style={S.styles.separator} />
+            <View style={styles.separator} />
             <ArtworkList
                 paymentType={selectedPaymentType}
             />
@@ -36,3 +35,20 @@ const CharacterSelectPage = () => {
 };
 
 export default CharacterSelectPage;
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        paddingLeft: 20,
+        paddingRight: 20,
+        position: 'relative',
+        backgroundColor: '#FFFFFF',
+    },
+    separator: {
+        width: "100%",
+        height: 1,
+        backgroundColor: '#F7F7F7',
+        alignSelf: 'stretch',
+        flexGrow: 0,
+    }
+});
