--- fourtoon-cookie/src/pages/CharacterSelectPage/TabsLayout/TabsLayout.tsx
@@ -1,52 +1,0 @@
-import React from 'react';
-import { Text, TouchableOpacity, View } from 'react-native';
-import { CharacterPaymentType } from '../../../types/character';
-
-import * as S from './TabsLayout.styled';
-import { useTranslationWithParentName } from '../../../hooks/locale';
-
-export interface TabsLayoutProps {
-    selectedPaymentType: CharacterPaymentType;
-    onSelectedPaymentTypeChange: (characterPaymentType: CharacterPaymentType) => void;
-}
-
-const TabsLayout = (props: TabsLayoutProps) => {
-    const { selectedPaymentType, onSelectedPaymentTypeChange, ...rest } = props;
-
-    const t = useTranslationWithParentName('pages.characterSelectPage.tabsLayout');
-    
-    return (
-        <View style={S.styles.headersContainer}>
-            {
-                [
-                    { type: CharacterPaymentType.FREE, label: t("free")},
-                    //{ type: CharacterPaymentType.PAID, label: t("paid") },
-                ].map((item) => (
-                    <Tab
-                        isActive={selectedPaymentType === item.type}
-                        label={item.label}
-                        onPress={() => onSelectedPaymentTypeChange(item.type)}
-                    />
-                ))
-            }
-        </View>
-    );
-};
-
-export default TabsLayout;
-
-interface TabProps {
-    isActive: boolean;
-    label: string;
-    onPress: () => void;
-}
-
-const Tab = (props: TabProps) => {
-    const { isActive, label, onPress, ...rest } = props;
-
-    return (
-        <TouchableOpacity onPress={onPress} style={[S.styles.tab, isActive && S.styles.selectedTab ]}>
-            <Text style={[S.styles.tabText, isActive && S.styles.activeTabText]}>{label}</Text>
-        </TouchableOpacity>
-    );
-};
